chore: remove legacy jest setup
