--- .github/workflows/ci.yml
@@ -1,95 +1,0 @@
-name: CI
-
-on:
-  push:
-  pull_request:
-  workflow_dispatch:
-    inputs:
-      ref:
-        description: 'Git ref to run against'
-        required: false
-        default: ''
-      phpunit_args:
-        description: 'Extra PHPUnit args (e.g. --filter MyTest)'
-        required: false
-        default: ''
-
-permissions:
-  contents: read
-
-concurrency:
-  group: ci-${{ github.ref }}
-  cancel-in-progress: true
-
-jobs:
-  php:
-    name: PHP (PHPUnit)
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-        with:
-          ref: ${{ github.event.inputs.ref || github.ref }}
-
-      - name: Setup PHP
-        uses: shivammathur/setup-php@v2
-        with:
-          php-version: '8.2'
-          coverage: none
-          tools: composer:v2
-
-      - name: Cache Composer
-        uses: actions/cache@v4
-        with:
-          path: |
-            ~/.composer/cache
-            vendor
-          key: ${{ runner.os }}-composer-${{ hashFiles('**/composer.lock') }}
-          restore-keys: |
-            ${{ runner.os }}-composer-
-
-      - name: Install PHP deps
-        run: composer install --no-interaction --prefer-dist --no-progress
-
-      - name: Set WP_PHPUNIT__TESTS_DIR
-        run: echo "WP_PHPUNIT__TESTS_DIR=vendor/wp-phpunit/wp-phpunit" >> $GITHUB_ENV
-
-      - name: Run PHPUnit
-        run: |
-          ARGS="${{ github.event.inputs.phpunit_args }}"
-          if [ -f phpunit.xml ] || [ -f phpunit.xml.dist ]; then
-            vendor/phpunit/phpunit/phpunit $ARGS
-          else
-            vendor/phpunit/phpunit/phpunit tests $ARGS
-          fi
-
-  js:
-    name: JS (Jest)
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-        with:
-          ref: ${{ github.event.inputs.ref || github.ref }}
-
-      - name: Setup Node
-        uses: actions/setup-node@v4
-        with:
-          node-version: 20
-
-      - name: Cache npm
-        uses: actions/cache@v4
-        with:
-          path: ~/.npm
-          key: ${{ runner.os }}-node-${{ hashFiles('**/package-lock.json') }}
-          restore-keys: |
-            ${{ runner.os }}-node-
-
-      - name: Install JS deps
-        run: npm ci
-
-      - name: Run Jest
-        env:
-          CI: 'true'
-        run: npm test -- --runInBand
-
