--- .github/workflows/codeql.yml
@@ -1,29 +1,0 @@
-name: CodeQL
-
-on:
-  push:
-  pull_request:
-  schedule:
-    - cron: '0 0 * * 0'
-  workflow_dispatch:
-    inputs:
-      ref:
-        description: 'Git ref to run against'
-        required: false
-        default: ''
-
-permissions:
-  contents: read
-  security-events: write
-
-jobs:
-  analyze:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          ref: ${{ github.event.inputs.ref || github.ref }}
-      - uses: github/codeql-action/init@v3
-        with:
-          languages: javascript,php
-      - uses: github/codeql-action/analyze@v3
