--- .github/workflows/docs-lint.yml
@@ -1,32 +1,0 @@
-name: Docs Lint
-on:
-  push:
-    paths:
-      - 'docs/**'
-      - '.github/workflows/docs-lint.yml'
-      - 'scripts/docs_validate.py'
-  pull_request:
-    paths:
-      - 'docs/**'
-      - '.github/workflows/docs-lint.yml'
-      - 'scripts/docs_validate.py'
-  workflow_dispatch:
-    inputs:
-      ref:
-        description: 'Git ref to run against'
-        required: false
-        default: ''
-
-jobs:
-  docs-check:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          ref: ${{ github.event.inputs.ref || github.ref }}
-      - name: Set up Python
-        uses: actions/setup-python@v5
-        with:
-          python-version: '3.x'
-      - name: Validate docs
-        run: python scripts/docs_validate.py
