--- .github/workflows/e2e.yml
@@ -1,51 +1,0 @@
-name: E2E
-
-on:
-  push:
-  pull_request:
-  workflow_dispatch:
-    inputs:
-      ref:
-        description: 'Git ref to run against'
-        required: false
-        default: ''
-
-jobs:
-  e2e:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          ref: ${{ github.event.inputs.ref || github.ref }}
-      - uses: actions/setup-node@v4
-        with:
-          node-version: 20
-      - uses: actions/cache@v4
-        with:
-          path: ~/.npm
-          key: ${{ hashFiles('**/package-lock.json') }}
-      - run: npm ci
-      - name: Start WordPress
-        run: npx @wordpress/env start
-      - name: Activate plugin
-        run: npx @wordpress/env run php wp plugin activate artpulse-management
-      - name: Seed users
-        run: |
-          npx @wordpress/env run php wp user create member member@example.com --role=member --user_pass=password
-          npx @wordpress/env run php wp user create artist artist@example.com --role=artist --user_pass=password
-          npx @wordpress/env run php wp user create organization org@example.com --role=organization --user_pass=password
-      - name: Cypress run
-        uses: cypress-io/github-action@v6
-        with:
-          command: npm run cypress:run
-      - name: Upload artifacts on failure
-        if: failure()
-        uses: actions/upload-artifact@v4
-        with:
-          name: cypress-artifacts
-          path: |
-            cypress/screenshots
-            cypress/videos
-      - name: Stop WordPress
-        if: always()
-        run: npx @wordpress/env stop
