+++ .github/workflows/orchestrate-all.yml
@@ -1,0 +1,59 @@
+name: orchestrate-all
+
+on:
+  workflow_dispatch:
+  schedule:
+    - cron: '0 1 * * *'
+
+permissions:
+  contents: read
+  actions: write
+
+jobs:
+  orchestrate:
+    runs-on: ubuntu-latest
+    timeout-minutes: 30
+    steps:
+      - uses: actions/checkout@v4
+      - name: Dispatch workflows
+        env:
+          GH_TOKEN: ${{ github.token }}
+          REF: ${{ github.ref }}
+        run: |
+          gh workflow list --json name | jq -r '.[] | select(.name != "orchestrate-all") | .name' > workflows.txt
+          while read wf; do
+            gh workflow run "$wf" --ref "$REF" || true
+          done < workflows.txt
+      - name: Await and summarize
+        env:
+          GH_TOKEN: ${{ github.token }}
+        run: |
+          echo '| Workflow | Conclusion | URL |' >> table.md
+          echo '|---|---|---|' >> table.md
+          while read wf; do
+            run_id=$(gh run list --workflow "$wf" --limit 1 --json databaseId -q '.[0].databaseId')
+            gh run watch $run_id || true
+            concl=$(gh run view $run_id --json conclusion -q .conclusion)
+            url="https://github.com/${{ github.repository }}/actions/runs/$run_id"
+            echo "| $wf | $concl | [logs]($url) |" >> table.md
+            gh run download $run_id -D artifacts/$wf || true
+          done < workflows.txt
+          cat table.md >> $GITHUB_STEP_SUMMARY
+          if [ -d artifacts ]; then
+            echo '\n### Artifacts' >> $GITHUB_STEP_SUMMARY
+            find artifacts -type f >> $GITHUB_STEP_SUMMARY
+          fi
+      - name: Upload logs
+        if: always()
+        uses: actions/upload-artifact@v4
+        with:
+          name: ${{ github.workflow }}-${{ github.job }}-logs
+          path: |
+            **/test-results/**/*.xml
+            **/logs/**
+          if-no-files-found: ignore
+      - name: Job summary
+        if: always()
+        run: |
+          echo "### ${{ github.workflow }} / ${{ github.job }}" >> $GITHUB_STEP_SUMMARY
+          echo "- status: ${{ job.status }}" >> $GITHUB_STEP_SUMMARY
