--- .github/workflows/phpunit.yml
@@ -1,79 +1,0 @@
-name: PHPUnit
-
-on:
-  pull_request:
-  push:
-    branches: [ main ]
-  workflow_dispatch:
-    inputs:
-      ref:
-        description: 'Git ref to run against'
-        required: false
-        default: ''
-
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    services:
-      mysql:
-        image: mysql:5.7
-        env:
-          MYSQL_ROOT_PASSWORD: root
-          MYSQL_DATABASE: wordpress_test
-        ports: ['3306:3306']
-        options: >-
-          --health-cmd="mysqladmin ping" --health-interval=10s --health-timeout=5s --health-retries=3
-
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          ref: ${{ github.event.inputs.ref || github.ref }}
-      - uses: shivammathur/setup-php@v2
-        with:
-          php-version: '8.3'
-      - run: composer install --prefer-dist --no-progress
-      - run: bin/install-wp-tests.sh wordpress_test root root 127.0.0.1 latest
-      - name: Run focused WP groups
-        run: vendor/bin/phpunit -c phpunit.wp.xml.dist --group admin --group restapi --group core --group engagement --group personalization
-        env:
-          WP_PHPUNIT__DIR: vendor/wp-phpunit/wp-phpunit
-          WP_TESTS_CONFIG_FILE_PATH: wp-tests-config.php
-          DB_NAME: wordpress_test
-          DB_USER: root
-          DB_PASSWORD: root
-          DB_HOST: 127.0.0.1
-          WP_CORE_DIR: /tmp/wordpress
-          AP_TEST_MODE: 1
-      - name: Run tests with coverage
-        run: tools/run-coverage.sh
-        env:
-          WP_PHPUNIT__DIR: vendor/wp-phpunit/wp-phpunit
-          WP_TESTS_CONFIG_FILE_PATH: wp-tests-config.php
-          DB_NAME: wordpress_test
-          DB_USER: root
-          DB_PASSWORD: root
-          DB_HOST: 127.0.0.1
-          WP_CORE_DIR: /tmp/wordpress
-          AP_TEST_MODE: 1
-      - name: Run REST API tests
-        id: rest_buckets
-        continue-on-error: true
-        run: composer run test:rest:buckets
-        env:
-          WP_PHPUNIT__DIR: vendor/wp-phpunit/wp-phpunit
-          WP_PHPUNIT__TESTS_CONFIG: tests/wp-tests-config.php
-          WP_TESTS_DB_NAME: wordpress_test
-          WP_TESTS_DB_USER: root
-          WP_TESTS_DB_PASSWORD: root
-          WP_TESTS_DB_HOST: 127.0.0.1
-          WP_CORE_DIR: /tmp/wordpress
-          AP_TEST_MODE: 1
-      - name: Upload REST buckets
-        if: always()
-        uses: actions/upload-artifact@v4
-        with:
-          name: rest-buckets
-          path: reports/rest-buckets.json
-      - name: Fail if REST buckets failed
-        if: steps.rest_buckets.outcome == 'failure'
-        run: exit 1
