--- .github/workflows/plugin-release.yml
@@ -1,58 +1,0 @@
-name: Plugin Release
-
-on:
-  push:
-    branches: [main]
-  pull_request:
-  workflow_dispatch:
-    inputs:
-      ref:
-        description: 'Git ref to run against'
-        required: false
-        default: ''
-
-env:
-  DB_NAME: wordpress_test
-  DB_USER: root
-  DB_PASSWORD: root
-  DB_HOST: 127.0.0.1
-  DB_CHARSET: utf8
-  DB_COLLATE: ''
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          ref: ${{ github.event.inputs.ref || github.ref }}
-      - uses: shivammathur/setup-php@v2
-        with:
-          php-version: '8.3'
-      - uses: actions/setup-node@v4
-        with:
-          node-version: 20
-      - name: Cache Composer
-        uses: actions/cache@v4
-        with:
-          path: |
-            ~/.composer/cache
-            vendor
-          key: composer-${{ runner.os }}-${{ hashFiles('composer.lock') }}
-      - name: Cache npm
-        uses: actions/cache@v4
-        with:
-          path: ~/.npm
-          key: npm-${{ runner.os }}-${{ hashFiles('package-lock.json') }}
-      - run: composer install --prefer-dist --no-progress
-      - run: npm ci
-      - run: npm run lint
-      - run: npm run test:js
-      - run: composer test
-      - run: composer sniff
-      - run: npm run build
-      - run: bash scripts/release.sh
-      - uses: actions/upload-artifact@v3
-        with:
-          name: plugin-zip
-          path: release/*.zip
