--- .github/workflows/static-checks.yml
@@ -1,65 +1,0 @@
-name: Static Checks
-
-on:
-  push:
-  pull_request:
-  workflow_dispatch:
-    inputs:
-      ref:
-        description: 'Git ref to run against'
-        required: false
-        default: ''
-
-jobs:
-  static:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          ref: ${{ github.event.inputs.ref || github.ref }}
-
-      - uses: shivammathur/setup-php@v2
-        with:
-          php-version: '8.3'
-          tools: composer
-
-      - name: Cache Composer dependencies
-        uses: actions/cache@v4
-        with:
-          path: ~/.composer/cache/files
-          key: composer-${{ runner.os }}-${{ hashFiles('**/composer.lock') }}
-          restore-keys: composer-${{ runner.os }}-
-
-      - name: Install dependencies
-        run: composer install --prefer-dist --no-progress
-
-      - name: PHP CodeSniffer
-        run: vendor/bin/phpcs --standard=phpcs.xml src --ignore=tests/TestHelpers/* -d memory_limit=512M
-
-      - name: PHPStan
-        run: |
-          set -e
-          sed '/phpstan-baseline.neon/d' phpstan.neon > phpstan-nobaseline.neon
-          vendor/bin/phpstan analyse -c phpstan-nobaseline.neon --error-format=json --no-progress --memory-limit=512M > phpstan.json
-          ERRORS=$(php -r 'echo json_decode(file_get_contents("phpstan.json"), true)["totals"]["errors"];')
-          BASELINE=$(grep -oP '^\s*count:\s*\K\d+' phpstan-baseline.neon | awk '{s+=$1} END {print s}')
-          echo "PHPStan errors: $ERRORS (baseline: $BASELINE)"
-          if [ "$ERRORS" -gt "$BASELINE" ]; then
-            echo "PHPStan errors exceed baseline"
-            exit 1
-          fi
-
-      - name: PHPUnit
-        run: |
-          mkdir -p build
-          vendor/bin/phpunit -c phpunit.unit.xml.dist --log-junit build/junit.xml
-        env:
-          WP_PHPUNIT__DIR: vendor/wp-phpunit/wp-phpunit
-          AP_TEST_MODE: '1'
-
-      - name: Upload JUnit report
-        if: always()
-        uses: actions/upload-artifact@v4
-        with:
-          name: phpunit-junit
-          path: build/junit.xml
