--- .github/workflows/tests.yml
@@ -1,207 +1,0 @@
-name: Tests
-
-on:
-  pull_request:
-  push:
-    branches: [main]
-    tags:
-      - '*'
-  schedule:
-    - cron: '0 2 * * *'
-  workflow_dispatch:
-
-jobs:
-  prepare:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-
-      - uses: shivammathur/setup-php@v2
-        with:
-          php-version: '8.2'
-          coverage: xdebug
-
-      - uses: actions/setup-node@v4
-        with:
-          node-version: 20
-
-      - run: composer install --no-interaction --no-progress --prefer-dist
-
-      - run: npm ci
-
-      - uses: actions/upload-artifact@v4
-        with:
-          name: deps
-          path: |
-            vendor
-            node_modules
-
-  lint-and-types:
-    runs-on: ubuntu-latest
-    needs: prepare
-    steps:
-      - uses: actions/checkout@v4
-
-      - uses: shivammathur/setup-php@v2
-        with:
-          php-version: '8.2'
-          coverage: xdebug
-
-      - uses: actions/setup-node@v4
-        with:
-          node-version: 20
-
-      - uses: actions/download-artifact@v4
-        with:
-          name: deps
-          path: .
-
-      - run: composer lint:php
-
-      - run: npm run lint
-
-      - run: npm run typecheck
-
-  phpunit:
-    runs-on: ubuntu-latest
-    needs: prepare
-    services:
-      mysql:
-        image: mysql:5.7
-        env:
-          MYSQL_ROOT_PASSWORD: root
-          MYSQL_DATABASE: wordpress_test
-        ports:
-          - 3306:3306
-        options: >-
-          --health-cmd="mysqladmin ping --silent"
-          --health-interval=10s
-          --health-timeout=5s
-          --health-retries=3
-    env:
-      DB_HOST: 127.0.0.1
-      DB_USER: root
-      DB_PASSWORD: root
-      DB_NAME: wordpress_test
-    steps:
-      - uses: actions/checkout@v4
-
-      - uses: shivammathur/setup-php@v2
-        with:
-          php-version: '8.2'
-          coverage: xdebug
-
-      - uses: actions/setup-node@v4
-        with:
-          node-version: 20
-
-      - uses: actions/download-artifact@v4
-        with:
-          name: deps
-          path: .
-
-      - run: npm run test:php
-
-      - uses: actions/upload-artifact@v4
-        with:
-          name: phpunit-coverage
-          path: build/coverage-phpunit*.xml
-
-  jest:
-    runs-on: ubuntu-latest
-    needs: prepare
-    steps:
-      - uses: actions/checkout@v4
-
-      - uses: actions/setup-node@v4
-        with:
-          node-version: 20
-
-      - uses: actions/download-artifact@v4
-        with:
-          name: deps
-          path: .
-
-      - run: npm run test:js -- --coverage
-
-      - uses: actions/upload-artifact@v4
-        with:
-          name: jest-coverage
-          path: coverage/lcov.info
-
-  cypress:
-    runs-on: ubuntu-latest
-    needs: prepare
-    steps:
-      - uses: actions/checkout@v4
-
-      - uses: shivammathur/setup-php@v2
-        with:
-          php-version: '8.2'
-          coverage: xdebug
-
-      - uses: actions/setup-node@v4
-        with:
-          node-version: 20
-
-      - uses: actions/download-artifact@v4
-        with:
-          name: deps
-          path: .
-
-      - uses: cypress-io/github-action@v6
-        with:
-          start: node scripts/spin-wp.js
-          command: npm run cypress:run
-
-      - uses: actions/upload-artifact@v4
-        with:
-          name: cypress-junit
-          path: build/e2e/cypress.xml
-
-  codex-checks:
-    runs-on: ubuntu-latest
-    needs: prepare
-    steps:
-      - uses: actions/checkout@v4
-
-      - uses: shivammathur/setup-php@v2
-        with:
-          php-version: '8.2'
-          coverage: xdebug
-
-      - uses: actions/setup-node@v4
-        with:
-          node-version: 20
-
-      - uses: actions/download-artifact@v4
-        with:
-          name: deps
-          path: .
-
-      - run: npm run codex:checks
-
-  coverage-merge:
-    runs-on: ubuntu-latest
-    needs:
-      - phpunit
-      - jest
-    steps:
-      - uses: actions/checkout@v4
-
-      - uses: actions/setup-node@v4
-        with:
-          node-version: 20
-
-      - uses: actions/download-artifact@v4
-        with:
-          name: phpunit-coverage
-          path: coverage
-
-      - uses: actions/download-artifact@v4
-        with:
-          name: jest-coverage
-          path: coverage
-
-      - run: node scripts/merge-coverage.js
-
